+++ lib/AppInfo/Application.php
@@ -73,6 +73,7 @@ class Application extends App implements IBootstrap {
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\EvaAudioChatProvider::class);
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\MultimodalChatWithToolsProvider::class);
         $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\ContextAgentAudioProvider::class);
+        $context->registerTaskProcessingProvider(\OCA\EvaAi\TaskProcessing\MultimodalContextAgentProvider::class);
         // Talk-Bot: reagiert auf BotInvokeEvent, wenn Nextcloud Talk installiert ist.
         $context->registerEventListener(\OCA\Talk\Events\BotInvokeEvent::class, \OCA\EvaAi\Listener\TalkBotListener::class);
         // GDPR-Erasure: Kontenloeschung raeumt alle eva_ai-Daten des Users ab (Issue #83).